\name{BullingerLeukemia}
\alias{BullingerLeukemia}
\docType{data}

\title{Leukemia data}

\description{Gene expression profiles of adult acute myeloid leukemia patients. Contains 116 samples and 1571 genes. The follow-up data also included.}

\usage{data(BullingerLeukemia)}

\format{An object of list class.}

\source{Bullinger,L. et al., (2004). "Use of gene-expression profiling to identify prognostic subclasses in adult acute myeloid leukemia", \emph{N Engl J Med.}, 350,  1605-1616.}


\note{The complete dataset is available at the Gene Expression Omnibus \url{www.ncbi.nlm.nih.gov/geo/}, accession number GSE425.}

\examples{
data(BullingerLeukemia)
names(BullingerLeukemia)
}

